test(passkey): clean up webauthn-adapter integration tests
Because:

* tests are duplicated across webauthn-adapter.spec.ts and .in.spec.ts.

This commit:

* removes webauthn-adapter.spec.ts
* adds two more tests for sign count rollback

Closes FXA-13462
